+++ tensorrt_llm/_torch/speculative/drafting_loops.py
@@ -15,7 +15,7 @@ from typing import Optional, final
 import torch
 
 from tensorrt_llm._torch.attention_backend.interface import AttentionMetadata
-from tensorrt_llm._torch.pyexecutor.sampler.sampling_utils import \
+from tensorrt_llm._torch.pyexecutor.sampler.ops.vanilla import \
     greedy_search_sampling_batch
 from tensorrt_llm._torch.speculative.eagle3 import Eagle3SpecMetadata
 from tensorrt_llm._torch.speculative.interface import SpecMetadata
